// components/shared/product/product-card.tsx
import React from "react";
import { Product } from "@/types/Product";
import Link from "next/link";
import Image from "next/image";
import { Card, CardContent, CardHeader } from "@/components/ui/card";
import { Car } from "lucide-react";
import ProductPrice from "./product-price";
import ProductList from "./product-list";

export default function ProductCard({product}:{product: Product}){
    return(
        <Card className="w-full max-w-sm">
            <CardHeader className="p-0 items-center">
            <Link href={`/product/${product.slug}`}>   
            <Image src={product.images[0]} alt={product.name} height={300} width={300} priority></Image>
            </Link>
            </CardHeader>

            <CardContent className="p-4 grid gap-4">
                <div className="text-s">{product.brand}</div>
                <Link href={`/product/${product.slug}`}>
                    <div className="text-sm font-medium">{product.name}</div>
                </Link>
                <div className="flex-between gap-4">
                    <p>{product.rating} Stars</p>
                    
                    {product.stock >0 ? <ProductPrice value={product.price}></ProductPrice> : <p className="text-destructive">Out of Stock</p>}
                    
                </div>
            </CardContent>
        </Card>
    )
}